Fixing store releated issues
