+++ features/apimgt/org.wso2.carbon.apimgt.store.feature/src/main/resources/devportal/source/src/app/components/Apis/Details/Credentials/Credentials.jsx
@@ -20,17 +20,21 @@ import PropTypes from 'prop-types';
 import { withStyles } from '@material-ui/core/styles';
 import { Typography } from '@material-ui/core';
 import Dialog from '@material-ui/core/Dialog';
+import Divider from '@material-ui/core/Divider';
 import Slide from '@material-ui/core/Slide';
+import Button from '@material-ui/core/Button';
 import Subscription from 'AppData/Subscription';
 import GenericDisplayDialog from 'AppComponents/Shared/GenericDisplayDialog';
 import Api from 'AppData/api';
 import Alert from 'AppComponents/Shared/Alert';
+import Paper from '@material-ui/core/Paper';
+import Grid from '@material-ui/core/Grid';
 import { FormattedMessage, injectIntl } from 'react-intl';
+import SubscribeToApi from 'AppComponents/Shared/AppsAndKeys/SubscribeToApi';
+import { ScopeValidation, resourceMethods, resourcePaths } from 'AppComponents/Shared/ScopeValidation';
 import { ApiContext } from '../ApiContext';
 import Wizard from './Wizard/Wizard';
 import SubscriptionTableRow from './SubscriptionTableRow';
-import SubscribeToApps from './SubscribeToApps';
-import SubscibeButtonPanel from './SubscibeButtonPanel';
 
 /**
  * @inheritdoc
@@ -39,16 +43,26 @@ import SubscibeButtonPanel from './SubscibeButtonPanel';
 const styles = theme => ({
     contentWrapper: {
         maxWidth: theme.custom.contentAreaWidth,
-        paddingLeft: theme.spacing.unit * 3,
-        paddingTop: theme.spacing.unit * 3,
+        paddingLeft: theme.spacing(3),
+        paddingTop: theme.spacing(3),
     },
     titleSub: {
-        cursor: 'pointer',
+        marginLeft: theme.spacing(2),
+        paddingTop: theme.spacing(2),
+        paddingBottom: theme.spacing(2),
+    },
+    generateCredentialWrapper: {
+        marginLeft: 0,
+        paddingTop: theme.spacing(2),
+        paddingBottom: theme.spacing(2),
     },
     tableMain: {
         width: '100%',
         borderCollapse: 'collapse',
-        marginTop: theme.spacing.unit * 3,
+        marginTop: theme.spacing(3),
+        marginLeft: theme.spacing(2),
+        marginRight: theme.spacing(1),
+
     },
     th: {
         color: theme.palette.getContrastText(theme.palette.background.default),
@@ -56,9 +70,9 @@ const styles = theme => ({
         borderTop: 'solid 1px ' + theme.palette.grey.A200,
         textAlign: 'left',
         fontSize: '11px',
-        paddingLeft: theme.spacing.unit,
-        paddingTop: theme.spacing.unit,
-        paddingBottom: theme.spacing.unit,
+        paddingLeft: theme.spacing(1),
+        paddingTop: theme.spacing(1),
+        paddingBottom: theme.spacing(1),
     },
     expansion: {
         background: 'transparent',
@@ -68,7 +82,7 @@ const styles = theme => ({
         alignItems: 'center',
     },
     subscribeRoot: {
-        paddingLeft: theme.spacing.unit * 2,
+        paddingLeft: theme.spacing(2),
     },
     activeLink: {
         background: theme.palette.grey.A100,
@@ -78,10 +92,27 @@ const styles = theme => ({
         color: theme.palette.getContrastText(theme.palette.background.paper),
     },
     toolbar: {
-        marginLeft: theme.spacing.unit * 2,
+        marginLeft: theme.spacing(2),
     },
     subscribeTitle: {
         flex: 1,
+    },
+    paper: {
+        marginLeft: theme.spacing(2),
+        padding: theme.spacing(2),
+    },
+    descWrapper: {
+        marginBottom: theme.spacing(2),
+    },
+    credentialBoxWrapper: {
+        paddingLeft: theme.spacing(2),
+    },
+    credentialBox: {
+        padding: theme.spacing(1),
+        border: 'solid 1px #ccc',
+        borderRadius: 5,
+        marginBottom: theme.spacing(2),
+        marginTop: theme.spacing(2),
     },
 });
 
@@ -125,16 +156,9 @@ class Credentials extends React.Component {
         } else {
             updateSubscriptionData(this.updateData);
         }
-        const { history: { location: { state, pathname }, replace } } = this.props;
-        if (state) {
-            const { openWizard } = state;
-            if (openWizard) {
-                this.setState({ wizardOn: true, openNew: true });
-                replace({
-                    pathname,
-                    state: {},
-                });
-            }
+        const { location: {pathname} } = this.props;
+        if( pathname.indexOf('wizard') !== -1){
+            this.setState({ wizardOn: true, openNew: true });
         }
     }
 
@@ -150,7 +174,7 @@ class Credentials extends React.Component {
             newSubscriptionRequest.applicationId = applicationsAvailable[0].value;
         }
         this.setState({ subscriptionRequest: newSubscriptionRequest, throttlingPolicyList });
-    }
+    };
 
     /**
      * @memberof Credentials
@@ -171,20 +195,25 @@ class Credentials extends React.Component {
      * @memberof Credentials
      */
     handleSubscribe = () => {
+        debugger;
         const { updateSubscriptionData, apiType } = this.context;
         const { subscriptionRequest } = this.state;
         const { intl } = this.props;
         const api = new Api();
         api.subscribe(
-            subscriptionRequest.apiId, subscriptionRequest.applicationId,
-            subscriptionRequest.throttlingPolicy.tierName, apiType,
+            subscriptionRequest.apiId,
+            subscriptionRequest.applicationId,
+            subscriptionRequest.throttlingPolicy,
+            apiType,
         )
             .then((response) => {
                 console.log('Subscription created successfully with ID : ' + response.body.subscriptionId);
-                Alert.info(intl.formatMessage({
-                    defaultMessage: 'Subscribed successfully',
-                    id: 'Apis.Details.Credentials.Credentials.subscribed.successfully',
-                }));
+                Alert.info(
+                    intl.formatMessage({
+                        defaultMessage: 'Subscribed successfully',
+                        id: 'Apis.Details.Credentials.Credentials.subscribed.successfully',
+                    }),
+                );
                 if (updateSubscriptionData) updateSubscriptionData();
                 this.setState({ openAvailable: false });
             })
@@ -224,7 +253,7 @@ class Credentials extends React.Component {
      */
     updateSubscriptionRequest = (subscriptionRequest) => {
         this.setState({ subscriptionRequest });
-    }
+    };
 
     /**
      *
@@ -239,20 +268,23 @@ class Credentials extends React.Component {
         promisedDelete.then((response) => {
             if (response.status !== 200) {
                 console.log(response);
-                Alert.info(intl.formatMessage({
-                    defaultMessage: 'Something went wrong while deleting the Subscription!',
-                    id: 'Apis.Details.Credentials.Credentials.something.went.wrong.with.subscription',
-                }));
+                Alert.info(
+                    intl.formatMessage({
+                        defaultMessage: 'Something went wrong while deleting the Subscription!',
+                        id: 'Apis.Details.Credentials.Credentials.something.went.wrong.with.subscription',
+                    }),
+                );
                 return;
             }
-            Alert.info(intl.formatMessage({
-                defaultMessage: 'Subscription deleted successfully!',
-                id: 'Apis.Details.Credentials.Credentials.subscription.deleted.successfully',
-            }));
+            Alert.info(
+                intl.formatMessage({
+                    defaultMessage: 'Subscription deleted successfully!',
+                    id: 'Apis.Details.Credentials.Credentials.subscription.deleted.successfully',
+                }),
+            );
             if (updateSubscriptionData) updateSubscriptionData();
         });
-    }
-
+    };
 
     /**
      * @inheritdoc
@@ -263,116 +295,208 @@ class Credentials extends React.Component {
             api, updateSubscriptionData, applicationsAvailable, subscribedApplications,
         } = this.context;
         const {
-            selectedKeyType, selectedAppId, wizardOn, openAvailable, subscriptionRequest,
-            throttlingPolicyList, openNew,
+            selectedKeyType,
+            selectedAppId,
+            wizardOn,
+            openAvailable,
+            subscriptionRequest,
+            throttlingPolicyList,
+            openNew,
         } = this.state;
         return (
-            <div className={classes.contentWrapper}>
-                <Typography onClick={this.handleExpandClick} variant='h4' className={classes.titleSub}>
-                    <FormattedMessage
-                        id='Apis.Details.Credentials.Credentials.api.credentials'
-                        defaultMessage='API Credentials'
-                    />
-                </Typography>
-                <Typography variant='body2' gutterBottom>
-                    <FormattedMessage
-                        id='Apis.Details.Credentials.Credentials.'
-                        defaultMessage={`API Credentials are grouped in to applications. An application is 
+            <Grid container>
+                <Grid item md={12} lg={9}>
+                    <Grid container spacing={5}>
+                        <Grid item md={12}>
+                            <Typography onClick={this.handleExpandClick} variant='h4' className={classes.titleSub}>
+                                <FormattedMessage
+                                    id='Apis.Details.Credentials.Credentials.api.credentials'
+                                    defaultMessage='API Credentials'
+                                />
+                            </Typography>
+                            <Paper elevation={0} className={classes.paper}>
+                                <Typography variant='body2' className={classes.descWrapper}>
+                                    <FormattedMessage
+                                        id='Apis.Details.Credentials.Credentials.'
+                                        defaultMessage={`API Credentials are grouped in to applications. An application is 
                         primarily used to decouple the consumer from the APIs. It allows you to Generate 
                         and use a single key for multiple APIs and subscribe multiple times to a single 
                         API with different SLA levels.`}
-                    />
-                </Typography>
-                {applicationsAvailable.length === 0 && subscribedApplications.length === 0 ? (
-                    !wizardOn && (
-                        <GenericDisplayDialog
-                            classes={classes}
-                            handleClick={this.startStopWizard}
-                            heading={intl.formatMessage({
-                                defaultMessage: 'Generate Credentials',
-                                id: 'Apis.Details.Credentials.Credentials.generate.credentials',
-                            })}
-                            caption={intl.formatMessage({
-                                defaultMessage: 'You need to generate credentials to access this API',
-                                id: 'Apis.Details.Credentials.Credentials.you.need.to'
-                                + '.generate.credentials.to.access.this.api',
-                            })}
-                            buttonText={intl.formatMessage({
-                                defaultMessage: 'GENERATE',
-                                id: 'Apis.Details.Credentials.Credentials.generate',
-                            })}
-                        />
-                    )
-                ) : (
-                    <React.Fragment>
-                        <SubscibeButtonPanel
-                            avalibleAppsLength={applicationsAvailable.length}
-                            subscribedAppsLength={subscribedApplications.length}
-                            handleClickToggle={this.handleClickToggle}
-                        />
-                        {/*
-                                ****************************
-                                Subscription List
-                                ***************************
-                                */}
-                        <table className={classes.tableMain}>
-                            <tr>
-                                <th className={classes.th}>Application Name</th>
-                                <th className={classes.th}>Throttling Tier</th>
-                                <th className={classes.th}>Application Status</th>
-                                <th className={classes.th} />
-                            </tr>
-                            {subscribedApplications.map((app, index) => (
-                                <SubscriptionTableRow
-                                    loadInfo={this.loadInfo}
-                                    handleSubscriptionDelete={this.handleSubscriptionDelete}
-                                    selectedAppId={selectedAppId}
-                                    updateSubscriptionData={updateSubscriptionData}
-                                    selectedKeyType={selectedKeyType}
-                                    app={app}
-                                    index={index}
-                                />
-                            ))}
-                        </table>
-                        {/*
-                                ****************************
-                                Subscribe to apps available
-                                ***************************
-                                */}
-                        {applicationsAvailable.length > 0 && (
-                            <SubscribeToApps
-                                api={api}
-                                openAvailable={openAvailable}
-                                handleClickToggle={this.handleClickToggle}
-                                Transition={Transition}
-                                applicationsAvailable={applicationsAvailable}
-                                handleSubscribe={this.handleSubscribe}
-                                subscriptionRequest={subscriptionRequest}
-                                updateSubscriptionRequest={this.updateSubscriptionRequest}
-                                throttlingPolicyList={throttlingPolicyList}
-                            />
-                        )}
-                        {/*
-                                ***************************************
-                                Subscribe with new Mode
-                                ***************************************
-                                */}
-                        <Dialog
-                            fullScreen
-                            open={openNew}
-                            onClose={() => this.handleClickToggle('openNew', updateSubscriptionData)}
-                            TransitionComponent={Transition}
-                        >
-                            <Wizard
-                                updateSubscriptionData={updateSubscriptionData}
-                                apiId={api.id}
-                                handleClickToggle={this.handleClickToggle}
-                                throttlingPolicyList={throttlingPolicyList}
-                            />
-                        </Dialog>
-                    </React.Fragment>
-                )}
-            </div>
+                                    />
+                                </Typography>
+
+                                {applicationsAvailable.length === 0 && subscribedApplications.length === 0 ? (
+                                    !wizardOn && (
+                                        <GenericDisplayDialog
+                                            classes={classes}
+                                            handleClick={this.startStopWizard}
+                                            heading={intl.formatMessage({
+                                                defaultMessage: 'Generate Credentials',
+                                                id: 'Apis.Details.Credentials.Credentials.generate.credentials',
+                                            })}
+                                            caption={intl.formatMessage({
+                                                defaultMessage: 'You need to generate credentials to access this API',
+                                                id:
+                                                    'Apis.Details.Credentials.Credentials.you.need.to'
+                                                    + '.generate.credentials.to.access.this.api',
+                                            })}
+                                            buttonText={intl.formatMessage({
+                                                defaultMessage: 'GENERATE',
+                                                id: 'Apis.Details.Credentials.Credentials.generate',
+                                            })}
+                                        />
+                                    )
+                                ) : (
+                                    <React.Fragment>
+                                        <div className={classes.generateCredentialWrapper}>
+                                            <ScopeValidation
+                                                resourcePath={resourcePaths.SUBSCRIPTIONS}
+                                                resourceMethod={resourceMethods.POST}
+                                            >
+                                                <Typography variant='h5'>
+                                                    <FormattedMessage
+                                                        id='Apis.Details.Credentials.Credentials.api.credentials.generate'
+                                                        defaultMessage='Generate Credentials'
+                                                    />
+                                                </Typography>
+                                                <div className={classes.credentialBoxWrapper}>
+                                                    <div className={classes.credentialBox}>
+                                                        <Typography variant='body2'>
+                                                            <FormattedMessage
+                                                                id='Apis.Details.Credentials.Credentials.api.credentials.with.wizard.message'
+                                                                defaultMessage={
+                                                                    'Use the Key Generation Wizard. Create a new application -> subscribe -> '
+                                                                    + ' generate keys and accesstoken to invoke this API.'
+                                                                }
+                                                            />
+                                                        </Typography>
+                                                        <Button
+                                                            variant='contained'
+                                                            size='large'
+                                                            color='primary'
+                                                            className={classes.buttonElm}
+                                                            onClick={() => this.handleClickToggle('openNew')}
+                                                        >
+                                                            <FormattedMessage
+                                                                id='Apis.Details.Credentials.SubscibeButtonPanel.subscribe.wizard'
+                                                                defaultMessage='Wizard'
+                                                            />
+                                                        </Button>
+                                                    </div>
+                                                    {applicationsAvailable.length > 0 && (
+                                                        <div className={classes.credentialBox}>
+                                                            <Typography variant='body2'>
+                                                                <FormattedMessage
+                                                                    id='Apis.Details.Credentials.Credentials.api.credentials.with.subscribe.message'
+                                                                    defaultMessage='Subscribe to an application and generate credentials'
+                                                                />
+                                                            </Typography>
+                                                            <SubscribeToApi
+                                                                applicationsAvailable={applicationsAvailable}
+                                                                subscriptionRequest={subscriptionRequest}
+                                                                throttlingPolicyList={throttlingPolicyList}
+                                                                updateSubscriptionRequest={
+                                                                    this.updateSubscriptionRequest
+                                                                }
+                                                                renderSmall
+                                                            />
+                                                            <Button
+                                                                variant='contained'
+                                                                size='large'
+                                                                color='primary'
+                                                                className={classes.buttonElm}
+                                                                onClick={() => this.handleSubscribe()}
+                                                            >
+                                                                <FormattedMessage
+                                                                    id='Apis.Details.Credentials.SubscibeButtonPanel.subscribe.btn'
+                                                                    defaultMessage='Subscribe'
+                                                                />
+                                                            </Button>
+                                                        </div>
+                                                    )}
+                                                </div>
+                                            </ScopeValidation>
+                                        </div>
+                                        {/*
+                                        ****************************
+                                        Subscription List
+                                        ***************************
+                                        */}
+                                        {subscribedApplications && subscribedApplications.length > 0 && (
+                                            <React.Fragment>
+                                                <Typography variant='h5'>
+                                                    <FormattedMessage
+                                                        id='Apis.Details.Credentials.Credentials.api.credentials.view.apps.title'
+                                                        defaultMessage='View Credentials'
+                                                    />
+                                                </Typography>
+                                                <Typography variant='body2'>
+                                                    <FormattedMessage
+                                                        id='Apis.Details.Credentials.Credentials.api.credentials.view.apps.title.sub'
+                                                        defaultMessage='( Subscribed Applications )'
+                                                    />
+                                                </Typography>
+                                                <table className={classes.tableMain}>
+                                                    <tr>
+                                                        <th className={classes.th}>
+                                                            <FormattedMessage
+                                                                id='Apis.Details.Credentials.Credentials.api.credentials.subscribed.apps.name'
+                                                                defaultMessage='Application Name'
+                                                            />
+                                                        </th>
+                                                        <th className={classes.th}>
+                                                            <FormattedMessage
+                                                                id='Apis.Details.Credentials.Credentials.api.credentials.subscribed.apps.tier'
+                                                                defaultMessage='Throttling Tier'
+                                                            />
+                                                        </th>
+                                                        <th className={classes.th}>
+                                                            <FormattedMessage
+                                                                id='Apis.Details.Credentials.Credentials.api.credentials.subscribed.apps.status'
+                                                                defaultMessage='Application Status'
+                                                            />
+                                                        </th>
+                                                        <th className={classes.th} />
+                                                    </tr>
+                                                    {subscribedApplications.map((app, index) => (
+                                                        <SubscriptionTableRow
+                                                            loadInfo={this.loadInfo}
+                                                            handleSubscriptionDelete={this.handleSubscriptionDelete}
+                                                            selectedAppId={selectedAppId}
+                                                            updateSubscriptionData={updateSubscriptionData}
+                                                            selectedKeyType={selectedKeyType}
+                                                            app={app}
+                                                            index={index}
+                                                        />
+                                                    ))}
+                                                </table>
+                                            </React.Fragment>
+                                        )}
+                                        {/*
+                                        ***************************************
+                                        Subscribe with new Mode
+                                        ***************************************
+                                        */}
+                                        <Dialog
+                                            fullScreen
+                                            open={openNew}
+                                            onClose={() => this.handleClickToggle('openNew', updateSubscriptionData)}
+                                            TransitionComponent={Transition}
+                                        >
+                                            <Wizard
+                                                updateSubscriptionData={updateSubscriptionData}
+                                                apiId={api.id}
+                                                handleClickToggle={this.handleClickToggle}
+                                                throttlingPolicyList={throttlingPolicyList}
+                                            />
+                                        </Dialog>
+                                    </React.Fragment>
+                                )}
+                            </Paper>
+                        </Grid>
+                    </Grid>
+                </Grid>
+            </Grid>
         );
     }
 }
