+++ features/apimgt/org.wso2.carbon.apimgt.store.feature/src/main/resources/devportal/source/src/app/components/Apis/Listing/NoApi.jsx
@@ -1,0 +1,38 @@
+import React from 'react';
+import { makeStyles, useTheme } from '@material-ui/core/styles';
+import Grid from '@material-ui/core/Grid';
+import Typography from '@material-ui/core/Typography';
+import { FormattedMessage } from 'react-intl';
+
+const useStyles = makeStyles(theme => ({
+    root: {
+        flexGrow: 1,
+    },
+    messageWrapper: {
+        marginTop: theme.spacing(4),
+        padding: theme.spacing(2),
+        textAlign: 'center',
+        color: theme.palette.text.secondary,
+    },
+}));
+
+export default function NoApi() {
+    const classes = useStyles();
+    const theme = useTheme();
+
+    return (
+        <div className={classes.root}>
+            <Grid container spacing={3}>
+                <Grid item xs={12} className={classes.messageWrapper}>
+                    <img src={theme.custom.noApiImage} className={classes.messageWrapper} />
+                    <Typography variant='h5' gutterBottom>
+                        <FormattedMessage id='Apis.Listing.NoApi.nodata.title' defaultMessage='No APIs Available' />
+                    </Typography>
+                    <Typography variant='subtitle1' gutterBottom>
+                        <FormattedMessage id='Apis.Listing.NoApi.nodata.content' defaultMessage='There is no apis to show you right now.' />
+                    </Typography>
+                </Grid>
+            </Grid>
+        </div>
+    );
+}
