+++ features/apimgt/org.wso2.carbon.apimgt.store.feature/src/main/resources/devportal/source/src/app/components/Base/Loading/Loading.jsx
@@ -21,9 +21,7 @@ import CircularProgress from '@material-ui/core/CircularProgress';
 
 const LoadingAnimation = (props) => {
     return (
-        <div>
-            <CircularProgress style={{ margin: 'auto', display: 'block' }} />
-        </div>
+        <div class="apim-dual-ring"></div>
     );
 };
